% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dri.R
\name{dri}
\alias{dri}
\title{Calculates DRI}
\usage{
dri(IC, R, Q)
}
\arguments{
\item{IC}{correlation matrix}

\item{R}{policy preference ranks}

\item{Q}{consideration ratings}
}
\value{
DRI matrix
}
\description{
Calculates DRI from a correlation matrix.
}
