% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dritools-package.R
\docType{package}
\name{dritools-package}
\alias{dritools}
\alias{dritools-package}
\title{dritools: Deliberative Reasoning Index Tools}
\description{
A collection of tools to analyze DRI, including integration with large language models.
}
\author{
\strong{Maintainer}: Gustavo Umbelino \email{gustavo.kreiaumbelino@zda.uzh.ch}

}
\keyword{internal}
