% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dritools.R
\docType{package}
\name{dritools}
\alias{dritools-package}
\alias{dritools}
\title{Deliberative Reasoning Index Tools}
\description{
A collection of tools to analyze DRI,
including integration with large language models.
}
\author{
Gustavo Umbelino \email{gustavo.kreiaumbelino@zda.uzh.ch}
}
