% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{format_statements}
\alias{format_statements}
\title{Format Statement}
\usage{
format_statements(statements)
}
\arguments{
\item{statements}{a vector of statements}
}
\value{
A string of formatted statements.
}
\description{
Formats a vector of statements as:
}
\details{
1. statement A
2. statement B
3. statement C
}
\examples{
format_statements(c("hi", "i", "am", "Gus"))
}
